Fix not decreasing the flow use_cnt reference counter in some cases from the app layer detection code. This caused some streams to never fully time out and thus clutter up the flow table and session pool.

+++ src/flow-util.c
@@ -56,7 +56,7 @@ void FlowFree(Flow *f)
  *  \param   proto  protocol which is needed to be mapped
  */
 
-int FlowGetProtoMapping(uint8_t proto) {
+uint8_t FlowGetProtoMapping(uint8_t proto) {
 
     switch (proto) {
         case IPPROTO_TCP:
